Move MoonModel to common, vastly simplify things.
