+++ src/client/components/Sidebar.vue
@@ -81,7 +81,7 @@ import {GameOptionsModel} from '@/models/GameOptionsModel';
 import GlobalParameterValue from '@/client/components/GlobalParameterValue.vue';
 import MoonGlobalParameterValue from '@/client/components/MoonGlobalParameterValue.vue';
 import {GlobalParameter} from '@/common/GlobalParameter';
-import {MoonModel} from '@/models/MoonModel';
+import {MoonModel} from '@/common/models/MoonModel';
 import PreferencesDialog from '@/client/components/PreferencesDialog.vue';
 
 export default Vue.extend({
