+++ src/client/components/moon/MoonSpace.vue
@@ -13,7 +13,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import {SpaceModel} from '@/models/SpaceModel';
+import {SpaceModel} from '@/common/models/SpaceModel';
 import {TileType} from '@/common/TileType';
 import Bonus from '@/client/components/Bonus.vue';
