+++ src/common/models/MoonModel.ts
@@ -1,0 +1,8 @@
+import {SpaceModel} from './SpaceModel';
+
+export interface MoonModel {
+  spaces: Array<SpaceModel>;
+  colonyRate: number;
+  miningRate: number;
+  logisticsRate: number;
+}
